Implement payment workflow with offers, coupons, stock management, Docker and Celery

+++ Backend/orders/services/admin.py
@@ -1,5 +1,5 @@
 from orders.models import Order
-
+from rest_framework.exceptions import ValidationError
 
 def update_order(*, order, validated_data):
     """
@@ -18,5 +18,23 @@ def delete_order(*, order):
     """
     Delete an order.
     """
+
+    if order.status != Order.OrderStatus.PENDING:
+            raise ValidationError(
+                {
+                    "message": (
+                        "Only pending orders can be deleted."
+                    )
+                }
+            )
+            
+    if hasattr(order, "payment"):
+        raise ValidationError(
+            {
+                "message": (
+                    "Orders with payments cannot be deleted."
+                )
+            }
+        )
 
     order.delete()
